requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
numpy
scikit-learn
tensorflow
h5py
joblib